break up index.js to multiple modules
