+++ scripts/package.json
@@ -4,12 +4,11 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "node flair/index.js"
+    "update": "node ./src/index.js"
   },
   "author": "",
   "license": "Apache-2.0",
   "dependencies": {
-    "@omisego/omg-js-childchain": "^4.0.1-1.0.0",
     "@omisego/omg-js-util": "^4.0.1-1.0.0",
     "dotenv": "^8.2.0",
     "lodash": "^4.17.19",
